+++ src/main/java/entities/Student.java
@@ -2,6 +2,7 @@ package entities;
 
 import lombok.*;
 import javax.persistence.*;
+import java.util.List;
 
 
 @AllArgsConstructor
@@ -14,10 +15,13 @@ public class Student {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String firstName;
     private String lastName;
     private String username;
     private String password;
+//
+//    @ManyToMany
+//    private List<Lesson> lessonList;
 }
